Anyone can propose a legislative change
First of all, the instructions tell you to "submit a pull request"; but
only developers know what that means.  Secondly, a good project commit
policy will give useful guidelines for writing a good commit message.
The guidelines here are roughly based on the git project's commit
policy.  That policy is written for software, and this one is for
legislation, but the intent is to give the person a checklist of
writing their reasoning down.